requirements.txt
Django==5.2
pillow==11.3.0
django-admin-sortable2==2.3.1